services/actions/app/requirements.txt
fastapi
uvicorn
requests
pydantic
python-jose[cryptography]
python-dotenv